remove duplicates (#186)
* refactor checkRestakeOperatorsManagerAccess

* update subgraphUrl logic
