+++ src/methods/vault/transactions/util/check/checkRestakeOperatorsManagerAccess.ts
@@ -1,7 +1,9 @@
 import type { CheckInput } from './types'
 
 
-const checkRestakeOperatorsManagerAccess = async ({ userAddress, vaultAddress, contracts }: CheckInput) => {
+type Action<Input, Output> = (props: Input) => Promise<Output>
+
+const checkAccess = async ({ userAddress, vaultAddress, contracts }: CheckInput) => {
   try {
     const restakingVaultContract = contracts.helpers.createVault({
       options: { isRestake: true },
@@ -17,6 +19,23 @@ const checkRestakeOperatorsManagerAccess = async ({ userAddress, vaultAddress, c
   }
   catch {}
 }
+
+const checkRestakeOperatorsManagerAccess = <Input, Output>(action: Action<Input, Output>) => (
+  async (values: Input) => {
+    try {
+      const result = await action(values)
+
+      return result
+    }
+    catch (actionError) {
+      return checkAccess(values as unknown as CheckInput)
+        .then(
+          () => Promise.reject(actionError),
+          (error) => Promise.reject(error)
+        )
+    }
+  }
+)
 
 
 export default checkRestakeOperatorsManagerAccess
